feat: Add SharePoint API integration and logging functionality
- Implemented SharePointApi class for managing folders and uploading files to SharePoint Online using app certificate authentication.
- Added Recurrence enum to define automation task recurrence frequencies.
- Enhanced base settings to include SharePoint configuration and recurrence settings.
- Developed BotRunnerLocal and BotRunnerMaestro classes to handle bot execution with logging, resource monitoring, and SharePoint integration.
- Created SQLDatabaseConnectorDict class for executing SQL queries and managing database connections.
- Added logging capabilities with LoggerConfig to manage log files and directory creation.
- Implemented SQL query for inserting execution logs into the database.
