+++ botcity/sql/query/insert_log.sql
