+++ botcity/utils/logger.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 from loguru import logger
 
-from core.config import settings
+from botcity.settings.base_settings import settings
 
 
 class LoggerConfig:
